% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-correlation.R
\name{apply_correlation_formatting}
\alias{apply_correlation_formatting}
\title{Apply correlation-specific formatting to a melliotab object}
\usage{
apply_correlation_formatting(mt)
}
\arguments{
\item{mt}{A melliotab object}
}
\value{
Modified melliotab object
}
\description{
Apply correlation-specific formatting to a melliotab object
}
\keyword{internal}
