% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-html.R
\name{build_table_label}
\alias{build_table_label}
\title{Build table label text (e.g., "Table 1" or "TABLE I")}
\usage{
build_table_label(mt)
}
\description{
Build table label text (e.g., "Table 1" or "TABLE I")
}
\keyword{internal}
