% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{build_title_html}
\alias{build_title_html}
\title{Build title HTML string for export (used by mt_copy)}
\usage{
build_title_html(x)
}
\description{
Build title HTML string for export (used by mt_copy)
}
\keyword{internal}
