% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capitalize_note_after_label}
\alias{capitalize_note_after_label}
\title{Capitalize prose note text after a period-ending note label}
\usage{
capitalize_note_after_label(note, label)
}
\description{
Capitalize prose note text after a period-ending note label
}
\keyword{internal}
