% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit.R
\name{compress_image_for_url}
\alias{compress_image_for_url}
\title{Compress image for URL transfer}
\usage{
compress_image_for_url(fig)
}
\arguments{
\item{fig}{Internal Mellio plot handoff object.}
}
\value{
Base64-encoded string or NULL if compression fails
}
\description{
Compress image for URL transfer
}
\keyword{internal}
