% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-columns.R
\name{detect_ci_cols}
\alias{detect_ci_cols}
\title{Detect CI columns}
\usage{
detect_ci_cols(headers)
}
\arguments{
\item{headers}{Character vector of column names}
}
\value{
Logical vector
}
\description{
Detect CI columns
}
\keyword{internal}
