% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-columns.R
\name{detect_leading_zero_cols}
\alias{detect_leading_zero_cols}
\title{Detect which columns should have leading zeros removed}
\usage{
detect_leading_zero_cols(headers)
}
\arguments{
\item{headers}{Character vector of column names}
}
\value{
Logical vector
}
\description{
Detect which columns should have leading zeros removed
}
\keyword{internal}
