% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{.compare_clean_terms}
\alias{.compare_clean_terms}
\title{Clean R term names to display labels}
\usage{
.compare_clean_terms(terms)
}
\description{
Clean R term names to display labels
}
\keyword{internal}
