% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-extractors.R
\name{.fa_fit_table}
\alias{.fa_fit_table}
\title{Build fit indices table from psych::fa}
\usage{
.fa_fit_table(x, style, title, number, note, source, decimals, p_decimals, ...)
}
\description{
Build fit indices table from psych::fa
}
\keyword{internal}
