% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-extractors.R
\name{.fa_loadings_table}
\alias{.fa_loadings_table}
\title{Build factor loadings table from psych::fa}
\usage{
.fa_loadings_table(
  x,
  style,
  title,
  number,
  note,
  source,
  cut,
  sort,
  decimals,
  p_decimals,
  ...
)
}
\description{
Build factor loadings table from psych::fa
}
\keyword{internal}
