% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-extractors.R
\name{.fa_variance_table}
\alias{.fa_variance_table}
\title{Build variance accounted for table from psych::fa}
\usage{
.fa_variance_table(
  x,
  style,
  title,
  number,
  note,
  source,
  decimals,
  p_decimals,
  ...
)
}
\description{
Build variance accounted for table from psych::fa
}
\keyword{internal}
