% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-extractors.R
\name{.fitdiff_comparison}
\alias{.fitdiff_comparison}
\title{Build chi-squared difference test table from FitDiff}
\usage{
.fitdiff_comparison(
  x,
  style,
  title,
  number,
  note,
  source,
  decimals,
  p_decimals,
  ...
)
}
\description{
Build chi-squared difference test table from FitDiff
}
\keyword{internal}
