% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-extractors.R
\name{.fitdiff_diff}
\alias{.fitdiff_diff}
\title{Build differences in fit indices table from FitDiff}
\usage{
.fitdiff_diff(x, style, title, number, note, source, decimals, p_decimals, ...)
}
\description{
Build differences in fit indices table from FitDiff
}
\keyword{internal}
