% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-extractors.R
\name{.fitdiff_fit}
\alias{.fitdiff_fit}
\title{Build model fit indices table from FitDiff}
\usage{
.fitdiff_fit(x, style, title, number, note, source, decimals, p_decimals, ...)
}
\description{
Build model fit indices table from FitDiff
}
\keyword{internal}
