% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-extractors.R
\name{.lavaan_fit_table}
\alias{.lavaan_fit_table}
\title{Build a fit indices table from a lavaan model}
\usage{
.lavaan_fit_table(
  x,
  style,
  title,
  number,
  note,
  source,
  decimals,
  p_decimals,
  ...
)
}
\description{
Build a fit indices table from a lavaan model
}
\keyword{internal}
