% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-numbers.R
\name{format_apa_number}
\alias{format_apa_number}
\title{Format a number according to APA rules}
\usage{
format_apa_number(
  value,
  col_type,
  decimals = 2L,
  p_decimals = 3L,
  remove_lz = TRUE
)
}
\arguments{
\item{value}{Character string containing the value}

\item{col_type}{Column type: "pvalue", "estimate", "statistic", "integer", "default"}

\item{decimals}{Number of decimal places for estimates/statistics}

\item{p_decimals}{Number of decimal places for p-values}

\item{remove_lz}{Whether to remove leading zeros (style-dependent)}
}
\value{
Formatted character string
}
\description{
Format a number according to APA rules
}
\keyword{internal}
