% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-numbers.R
\name{format_business_number}
\alias{format_business_number}
\title{Format a number with thousands separators and parenthetical negatives}
\usage{
format_business_number(value, use_parenthetical = TRUE)
}
\arguments{
\item{value}{Character string containing the number}

\item{use_parenthetical}{Whether to use parentheses for negatives}
}
\value{
Formatted character string
}
\description{
Format a number with thousands separators and parenthetical negatives
}
\keyword{internal}
