% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-numbers.R
\name{format_ci_value}
\alias{format_ci_value}
\title{Format a confidence interval value}
\usage{
format_ci_value(value, decimals = 2L)
}
\arguments{
\item{value}{Character string like "[0.12, 0.45]" or "0.12, 0.45"}

\item{decimals}{Number of decimal places}
}
\value{
Formatted CI string "[ll, ul]"
}
\description{
Format a confidence interval value
}
\keyword{internal}
