% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melliotab.R
\name{format_table_data}
\alias{format_table_data}
\title{Format table data according to APA/style rules}
\usage{
format_table_data(
  data,
  col_types,
  lz_cols,
  decimals,
  p_decimals,
  remove_lz,
  style_config
)
}
\description{
Format table data according to APA/style rules
}
\keyword{internal}
