% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-numbers.R
\name{get_sig_stars}
\alias{get_sig_stars}
\title{Get significance stars for a p-value}
\usage{
get_sig_stars(p_value)
}
\arguments{
\item{p_value}{Character string or numeric p-value}
}
\value{
Character string: "", "\emph{", "\strong{", or "}}"
}
\description{
Get significance stars for a p-value
}
\keyword{internal}
