% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-correlation.R
\name{is_correlation_matrix}
\alias{is_correlation_matrix}
\title{Check if a matrix looks like a correlation matrix}
\usage{
is_correlation_matrix(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
Logical
}
\description{
Check if a matrix looks like a correlation matrix
}
\keyword{internal}
