% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-columns.R
\name{is_count_header}
\alias{is_count_header}
\title{Check whether a header denotes a whole-number count column}
\usage{
is_count_header(header)
}
\description{
Check whether a header denotes a whole-number count column
}
\keyword{internal}
