% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-columns.R
\name{is_p_value_header}
\alias{is_p_value_header}
\title{Check whether a header denotes a p-value column}
\usage{
is_p_value_header(header)
}
\description{
Check whether a header denotes a p-value column
}
\keyword{internal}
