% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-columns.R
\name{is_stat_symbol}
\alias{is_stat_symbol}
\title{Check if a header is a statistical symbol that should be italicized}
\usage{
is_stat_symbol(header)
}
\arguments{
\item{header}{Column header string}
}
\value{
Logical
}
\description{
Check if a header is a statistical symbol that should be italicized
}
\keyword{internal}
