% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-columns.R
\name{is_text_label_header}
\alias{is_text_label_header}
\title{Check whether a header denotes a label column}
\usage{
is_text_label_header(header)
}
\description{
Check whether a header denotes a label column
}
\keyword{internal}
