% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{knit_print.melliotab}
\alias{knit_print.melliotab}
\title{Knit print method for R Markdown / Quarto}
\usage{
knit_print.melliotab(x, ...)
}
\arguments{
\item{x}{A melliotab object}

\item{...}{Additional arguments}
}
\value{
knitr output
}
\description{
Automatically detects the output format and uses the appropriate backend.
}
