% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-columns.R
\name{match_header}
\alias{match_header}
\title{Match a header against a pattern set (case-insensitive)}
\usage{
match_header(header, pattern_set)
}
\description{
Match a header against a pattern set (case-insensitive)
}
\keyword{internal}
