% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{mellio_copy_script}
\alias{mellio_copy_script}
\title{Generate JavaScript for Cmd+C / Ctrl+C clipboard copy in Viewer}
\usage{
mellio_copy_script(figure = FALSE)
}
\arguments{
\item{figure}{If TRUE, swaps img src with base64 before copying}
}
\value{
HTML script tag string
}
\description{
Generate JavaScript for Cmd+C / Ctrl+C clipboard copy in Viewer
}
\keyword{internal}
