% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-html.R
\name{mt_as_gt}
\alias{mt_as_gt}
\title{Convert a melliotab object to a gt table}
\usage{
mt_as_gt(x)
}
\arguments{
\item{x}{A melliotab object}
}
\value{
A gt table object
}
\description{
Convert a melliotab object to a gt table
}
