% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-latex.R
\name{mt_as_latex}
\alias{mt_as_latex}
\title{Convert a melliotab object to LaTeX}
\usage{
mt_as_latex(x)
}
\arguments{
\item{x}{A melliotab object}
}
\value{
Character string of LaTeX code
}
\description{
Convert a melliotab object to LaTeX
}
