% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-markdown.R
\name{mt_as_markdown}
\alias{mt_as_markdown}
\title{Convert a melliotab object to Markdown}
\usage{
mt_as_markdown(x)
}
\arguments{
\item{x}{A melliotab object}
}
\value{
Character string of Markdown table
}
\description{
Convert a melliotab object to Markdown
}
