% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{mt_decimals}
\alias{mt_decimals}
\title{Set decimal places}
\usage{
mt_decimals(x, decimals = 2L, p_decimals = 3L)
}
\arguments{
\item{x}{A melliotab object}

\item{decimals}{Decimal places for estimates/statistics}

\item{p_decimals}{Decimal places for p-values}
}
\value{
Modified melliotab object
}
\description{
Set decimal places
}
