% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{mt_diagonal}
\alias{mt_diagonal}
\title{Set correlation matrix display options}
\usage{
mt_diagonal(
  x,
  mode = c("dash", "one", "blank"),
  triangle = c("all", "lower", "upper")
)
}
\arguments{
\item{x}{A melliotab object}

\item{mode}{Diagonal display: "dash", "one", "blank"}

\item{triangle}{Triangle to show: "all", "lower", "upper"}
}
\value{
Modified melliotab object
}
\description{
Set correlation matrix display options
}
