% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{mt_format_ci}
\alias{mt_format_ci}
\title{Format confidence intervals}
\usage{
mt_format_ci(x, decimals = NULL, bracket = TRUE)
}
\arguments{
\item{x}{A melliotab object}

\item{decimals}{Decimal places for CI values (NULL uses table default)}

\item{bracket}{Whether to use bracket notation}
}
\value{
Modified melliotab object
}
\description{
Format confidence intervals
}
