% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{mt_indent}
\alias{mt_indent}
\title{Indent row labels}
\usage{
mt_indent(x, rows, level = 1L)
}
\arguments{
\item{x}{A melliotab object}

\item{rows}{Row indices to indent}

\item{level}{Indentation level (1-3)}
}
\value{
Modified melliotab object
}
\description{
Indent row labels
}
