% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{mt_note}
\alias{mt_note}
\title{Set table note}
\usage{
mt_note(x, note)
}
\arguments{
\item{x}{A melliotab object}

\item{note}{Note text}
}
\value{
Modified melliotab object
}
\description{
Set table note
}
