% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{mt_number}
\alias{mt_number}
\title{Set table number}
\usage{
mt_number(x, number)
}
\arguments{
\item{x}{A melliotab object}

\item{number}{Table number}
}
\value{
Modified melliotab object
}
\description{
Set table number
}
