% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{mt_remove_leading_zeros}
\alias{mt_remove_leading_zeros}
\title{Control leading zero removal}
\usage{
mt_remove_leading_zeros(x, enabled = TRUE)
}
\arguments{
\item{x}{A melliotab object}

\item{enabled}{Whether to remove leading zeros}
}
\value{
Modified melliotab object
}
\description{
Control leading zero removal
}
