% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{mt_save}
\alias{mt_save}
\title{Save a melliotab table to file}
\usage{
mt_save(x, filename, ...)
}
\arguments{
\item{x}{A melliotab object}

\item{filename}{Output file path. Supported extensions: .html, .tex, .md}

\item{...}{Additional arguments passed to format-specific functions}
}
\value{
Invisible file path
}
\description{
Auto-detects the format from the file extension.
}
\examples{
\dontrun{
model <- lm(mpg ~ wt + hp, data = mtcars)
tbl <- melliotab(model, style = "apa7", title = "Results")

mt_save(tbl, "table.html")
mt_save(tbl, "table.tex")
mt_save(tbl, "table.md")
}
}
