% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{mt_section_title}
\alias{mt_section_title}
\title{Add a section title row}
\usage{
mt_section_title(x, label, before = NULL, after = NULL)
}
\arguments{
\item{x}{A melliotab object}

\item{label}{Section title text}

\item{before}{Row index to insert before}

\item{after}{Row index to insert after}
}
\value{
Modified melliotab object
}
\description{
Add a section title row
}
