% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{mt_set_style}
\alias{mt_set_style}
\title{Change the citation style}
\usage{
mt_set_style(x, style = c("apa7", "ieee"))
}
\arguments{
\item{x}{A melliotab object}

\item{style}{Style name: "apa7" or "ieee"}
}
\value{
Modified melliotab object
}
\description{
Change the citation style
}
