% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{mt_sig_stars}
\alias{mt_sig_stars}
\title{Add significance stars}
\usage{
mt_sig_stars(
  x,
  target = "auto",
  remove_p = TRUE,
  levels = c(`*` = 0.05, `**` = 0.01, `***` = 0.001)
)
}
\arguments{
\item{x}{A melliotab object}

\item{target}{Column to add stars to ("auto" to detect)}

\item{remove_p}{Whether to remove the p-value column}

\item{levels}{Named vector of significance thresholds}
}
\value{
Modified melliotab object
}
\description{
Add significance stars
}
