% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{mt_simplify_headers}
\alias{mt_simplify_headers}
\title{Simplify SPSS verbose headers}
\usage{
mt_simplify_headers(x)
}
\arguments{
\item{x}{A melliotab object}
}
\value{
Modified melliotab object
}
\description{
Simplify SPSS verbose headers
}
