% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{mt_source}
\alias{mt_source}
\title{Set table source}
\usage{
mt_source(x, source)
}
\arguments{
\item{x}{A melliotab object}

\item{source}{Source text}
}
\value{
Modified melliotab object
}
\description{
Set table source
}
