% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{mt_spanner}
\alias{mt_spanner}
\title{Add a spanning column header}
\usage{
mt_spanner(x, label, columns, level = 1L)
}
\arguments{
\item{x}{A melliotab object}

\item{label}{Spanner label text}

\item{columns}{Column indices or names covered by the spanner}

\item{level}{Spanner level (1 = closest to data)}
}
\value{
Modified melliotab object
}
\description{
Add a spanning column header
}
