% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{mt_title}
\alias{mt_title}
\title{Set table title}
\usage{
mt_title(x, title)
}
\arguments{
\item{x}{A melliotab object}

\item{title}{Title text}
}
\value{
Modified melliotab object
}
\description{
Set table title
}
