% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.melliotab}
\alias{print.melliotab}
\title{Print a melliotab object}
\usage{
\method{print}{melliotab}(x, ...)
}
\arguments{
\item{x}{A melliotab object}

\item{...}{Ignored}
}
\value{
Invisible melliotab object
}
\description{
Renders the table in the RStudio Viewer pane (or browser).
}
