% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-numbers.R
\name{remove_leading_zero}
\alias{remove_leading_zero}
\title{Remove leading zero from a bounded statistic value}
\usage{
remove_leading_zero(value)
}
\arguments{
\item{value}{Character string}
}
\value{
Character string with leading zero removed if applicable
}
\description{
Remove leading zero from a bounded statistic value
}
\keyword{internal}
