% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header-maps.R
\name{rename_to_apa}
\alias{rename_to_apa}
\title{Rename R/broom column names to APA standard abbreviations}
\usage{
rename_to_apa(headers)
}
\arguments{
\item{headers}{Character vector of column names}
}
\value{
Character vector with renamed headers
}
\description{
Rename R/broom column names to APA standard abbreviations
}
\keyword{internal}
