% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header-maps.R
\name{simplify_spss_header}
\alias{simplify_spss_header}
\title{Simplify an SPSS verbose header to APA abbreviation}
\usage{
simplify_spss_header(header)
}
\arguments{
\item{header}{Character string}
}
\value{
Simplified header string
}
\description{
Simplify an SPSS verbose header to APA abbreviation
}
\keyword{internal}
