% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit.R
\name{table_to_tsv}
\alias{table_to_tsv}
\title{Convert a data frame to tab-delimited text}
\usage{
table_to_tsv(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A single character string with tab-separated values
}
\description{
Convert a data frame to tab-delimited text
}
\keyword{internal}
