% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unicode_to_latex}
\alias{unicode_to_latex}
\title{Convert Unicode statistical symbols to LaTeX equivalents}
\usage{
unicode_to_latex(text)
}
\description{
Convert Unicode statistical symbols to LaTeX equivalents
}
\keyword{internal}
